The Scalar code is now broken down into their own implementations for easier maintenance
